#85567 query for items missing Release attribute within folder
Currently fully functional except for the recursive folder query, which
fails to return the same results as the browser. For now we are just
using provided top-level parent folders as our source, doing a query
within those to get the artifacts lacking Release: Deposits and
generating a CSV of the format Artifact ID,Release so that we can update
things via DNG's CSV importer.