#113396 created BaseQueryHandler for RM, errors in this version
Issue was from reusing CCM queryhandler classes for the RestartManager;
this cannot work as those are custom to CCM (though currently they are
named CLM so it's confusing). RM implementation created in RM scanned
packages so it all autowires nice.

Also commented out deprecated values in TranslateRmBaseArtifactsToADO
because it was getting confusing. There seems to be some redundancy in
the driver classes vs the QueryHandler controllers.  Maybe that is
normnal. Trying to comment code more.

Current errors in autowiring, working on resolving next Monday.