#114907 fixed restartmanager result filtering for DNG, update username
had a bit too much commented out and were missing variables, some name
issues, and the allFilter was targeting the wrong part of the tree.
Also: set svc-migration user to replace Eric's user as it should be a
fixed and sharable password.  This is updated in a few but not all
runconfigs.