Created set of sample batch files to run the 'full' migration
NOT COMPLETE - Requires input from Robert and Eric

The theory is, all parameters are set in orchestration_launcher for say,
dev or prod.  This calls orchestration master which starts the CCM, QM,
and RM migrations simultaneously, waiting for the RM migration to end
before moving on (eventually doing modules and linking).

If the parameters look good we should remove the 'echo' from the actual
run files; I put that in for now so that we do not accidentally run the
migration scripts, but can observe the output of the script.

#182974 created as a result of migration discrepancies from slight
differences in batch parameters