#406513 link processing phase added
Modified from the way other systems process links. For example, we do
not read the mapping; there are dozens of link mapping types out there,
and we COULD map them all manually to "Related" but that seems
pointless.  If there is a need to have a handful be some alternate
mapping, I have some ideas on taking care of that.

We ignore anything that isn't an RM-RM map because they will be taken
care of from the ccm/qm migration side.